Typo: StartingDeadlineSeconds should be int32

+++ docs/book/src/cronjob-tutorial/testdata/project/api/v1/cronjob_types.go
@@ -67,7 +67,7 @@ type CronJobSpec struct {
 	// Optional deadline in seconds for starting the job if it misses scheduled
 	// time for any reason.  Missed jobs executions will be counted as failed ones.
 	// +optional
-	StartingDeadlineSeconds *int64 `json:"startingDeadlineSeconds,omitempty"`
+	StartingDeadlineSeconds *int32 `json:"startingDeadlineSeconds,omitempty"`
 
 	// Specifies how to treat concurrent executions of a Job.
 	// Valid values are:
